poll_providers: persist token in Django cache
Indeed we were caching the token for the life of the polling session,
but as a cron job, a new session was started every minute or so...

The token is considered as a password and is encrypted. Not with
Django's secret key, the required key is a particular format.

Host applications are free to choose the cache of their choice.