provider_poller: support alternative time field
As the reworked comment at the top of the diff says, polling by
timestamp brings the issue of ignoring old telemetries saved in the
provider database after we already polled that point in time.

Better order by "saved_at" (as we call it in EventRecord) if the
provider API supports it.

When we activate that param in the API configuration, we also need to
reset the "last_start_time_polled" field.