provider_poller: split into different implementations
But reuse as much code as possible.

The goal is to avoid branching in the code to support multiple versions,
and regressions in other versions when we adjust the code for another
one.

Most of the benefit can be seen in forking the tests to covers all cases
with both versions.

Closes: SMP-1691